# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Minha API",
    "description": "Projeto de Gerenciamento de Atividades"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
...
    "/activity/": {
      "get": {
        "tags": [
          "Activity"
        ],
        "summary": "Retorna todos as atividades",
        "description": "Endpoint que busca todas as atividades",
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "post": {
        "tags": [
          "Activity"
        ],
        "summary": "Cria uma nova atividade",
        "description": "Endpoint que cria uma nova atividade",
        "responses": {
          "default": {
            "description": ""
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "description": {
                    "example": "any"
                  },
                  "userId": {
                    "example": "any"
                  },
                  "categoryId": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/activity/{id}": {
      "put": {
        "tags": [
          "Activity"
        ],
        "summary": "Atualiza uma atividade",
        "description": "Endpoint que altera dados de uma atividade cadastrada",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "Activity"
        ],
        "summary": "Remove uma atividade",
        "description": "Endpoint que remove uma atividade cadastrada",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}